Relocate project README and enhance package documentation
The README.md file for 'Frank.PulseFlow' was moved from within the project to the root of the repository. This was done to improve project visibility. Further enhancements were made to the README, including a detailed breakdown of key concepts involved in the lightweight, high-performance messaging system, its key features, use cases and installation instructions. Miscellaneous changes included the addition of supporting files like CONTRIBUTING.md and MAINTAINERS.md, renaming a test class to reflect its purpose better, and adding comment-based documentation to various classes and interfaces for a better understanding of their functionality.
