+++ Frank.PulseFlow.Sample/Directory.Build.props
@@ -1,0 +1,6 @@
+<Project>
+  <PropertyGroup>
+    <Nullable>enable</Nullable>
+    <AvaloniaVersion>11.0.6</AvaloniaVersion>
+  </PropertyGroup>
+</Project>
