+++ Frank.PulseFlow.Tests.Cli/Program.cs
@@ -1,6 +1,4 @@
-﻿// See https://aka.ms/new-console-template for more information
-
-using Frank.PulseFlow;
+﻿using Frank.PulseFlow;
 using Frank.PulseFlow.Tests.Cli;
 
 IHostBuilder builder = Host.CreateDefaultBuilder();
@@ -9,15 +7,11 @@ builder.ConfigureServices((context, services) =>
 {
     services.AddPulseFlow(messagingBuilder =>
     {
-        messagingBuilder.AddFlow<TextPulseFlow>();
+        messagingBuilder.AddFlow<TextFlow>();
     });
     services.AddHostedService<TestingService>();
 });
 
 IHost app = builder.Build();
 
-
 await app.RunAsync();
-
-
-//Extension Method
