+++ Frank.PulseFlow.sln
@@ -22,6 +22,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		restore.ps1 = restore.ps1
 		test.ps1 = test.ps1
 		pack.ps1 = pack.ps1
+		CONTRIBUTING.md = CONTRIBUTING.md
+		icon.png = icon.png
+		README.md = README.md
+		.github\workflows\main.yml = .github\workflows\main.yml
+		MAINTAINERS.md = MAINTAINERS.md
+		STYLE.md = STYLE.md
 	EndProjectSection
 EndProject
 Global
