+++ Frank.PulseFlow/Frank.PulseFlow.csproj
@@ -7,12 +7,6 @@
         <Description>PulseFlow uses Channel -mechanism for internal messaging</Description>
         <PackageTags>Pulse,Flow,PulseFlow,Frank,Frank Haugen,Messaging,In-Memory</PackageTags>
         
-        <PackageReleaseNotes>
-            2023-12-14:
-                1.0 Initial release of PulseFlow
-                1.1 Added support for .NET 7.0 and .NET 8.0
-        </PackageReleaseNotes>
-        
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <NoWarn>$(NoWarn);1591</NoWarn>
 
@@ -23,7 +17,7 @@
         <Authors>Frank R. Haugen</Authors>
         <PublisherName>Frank R. Haugen</PublisherName>
         <Copyright>Copyright (c) 2023 Frank R. Haugen</Copyright>
-        <PackageReadmeFile>readme.md</PackageReadmeFile>
+        <PackageReadmeFile>README.md</PackageReadmeFile>
         <PackageIcon>icon.png</PackageIcon>
         <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
         <PackageLicenseExpression>MIT</PackageLicenseExpression>
@@ -39,8 +33,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <None Include="readme.md" Pack="true" PackagePath="\"/>
-        <None Include="icon.png" Pack="true" PackagePath="\"/>
+        <None Include="../README.md" Pack="true" PackagePath="\"/>
+        <None Include="../icon.png" Pack="true" PackagePath="\"/>
     </ItemGroup>
 
     <ItemGroup>
