+++ Frank.PulseFlow/Internal/Conduit.cs
@@ -1,5 +1,8 @@
 namespace Frank.PulseFlow.Internal;
 
+/// <summary>
+/// A wrapper around a <see cref="Channel{T}"/> that sends and receives <see cref="IPulse"/> objects.
+/// </summary>
 internal class Conduit : IConduit
 {
     private readonly IChannel _messageChannel;
