+++ Frank.PulseFlow/Internal/IChannel.cs
@@ -1,5 +1,8 @@
 namespace Frank.PulseFlow.Internal;
 
+/// <summary>
+/// Represents a channel for sending and receiving pulses.
+/// </summary>
 internal interface IChannel
 {
     Task SendAsync(IPulse message);
