--- Frank.PulseFlow/README.md
@@ -1,125 +1,0 @@
-# PulseFlow Local Messaging
-
-PulseFlow Local Messaging is a lightweight, high-performance messaging system that enables seamless communication
-
-## Table of Contents
-
-- [Overview](#overview)
-  - [Key Features](#key-features)
-  - [Illustration](#illustration)
-  - [Use Cases](#use-cases)
-- [Getting Started](#getting-started)
-  - [Installation](#installation)
-- [Concepts](#concepts)
-  - [Nexus](#nexus)
-  - [Conduit](#conduit)
-  - [Pulse](#pulse)
-  - [PulseFlow](#pulseflow)
-
-## Overview
-
-PulseFlow Local Messaging is a lightweight, high-performance messaging system that enables seamless communication
-between different parts of an application. It's designed to be simple, flexible, and scalable, allowing for easy
-integration into any system architecture.
-
-### Key Features
-
-- **Lightweight**: PulseFlow is a lightweight messaging system, with a small footprint and minimal resource
-  requirements.
-- **High Performance**: It's designed for high performance and scalability, capable of handling a vast volume of
-  messages simultaneously.
-- **Asynchronous Communication**: PulseFlow supports asynchronous data flow, allowing for non-blocking and concurrent
-  message transmissions.
-- **Flexible and Adaptable**: It's flexible and adaptable, capable of handling various types of messages and adapting
-  its processing logic based on the nature and requirements of each message.
-- **Simple and Easy to Use**: PulseFlow is simple and easy to use, with a straightforward API and minimal configuration
-  requirements.
-
-### Illustration
-
-```mermaid
-graph TD
-    Pulse[Pulse: Data/Messages] -->|transmitted via| Conduit[Conduit: Message Pathway]
-    Conduit -->|delivered to| Nexus[Nexus: Processing Service]
-    Nexus -->|processed by| PulseFlow[PulseFlow: Message Handler]
-    PulseFlow -->|manipulates| Pulse
-```
-
-In this Mermaid diagram:
-- **Pulse** is shown as the starting point for data/messages.
-- **Conduit** is represented as the pathway for transmitting messages.
-- **Nexus** is the central processing service.
-- **PulseFlow** is depicted as handling and manipulating the messages.
-
-When you include this in a GitHub Markdown file, GitHub will render the Mermaid diagram as a visual graph. Remember to remove the extra backticks (```) in the beginning and end when adding this to your README.
-
-### Use Cases
-
-PulseFlow is a general-purpose messaging system that can be used in a wide variety of applications. It's particularly
-useful in scenarios where there's a need for asynchronous communication between different parts of the system.
-
-## Getting Started
-
-This section provides a quick guide on how to get started with PulseFlow Local Messaging.
-
-### Installation
-
-PulseFlow is available as a NuGet package, which can be installed using the following command:
-
-```bash
-dotnet add package Frank.PulseFlow
-```
-
-## Concepts
-
-This section provides an in-depth explanation of the key concepts and components within the system: Nexus, Conduit,
-Pulse, and PulseFlow. Understanding these concepts is crucial for grasping how the system operates and interacts with
-data.
-
-### Nexus
-
-The **Nexus** is the central hub of our messaging system, analogous to a neural network's core. It serves as the primary
-processing service, where all data messages, or 'Pulses', are received, interpreted, and routed to their respective
-destinations.
-
-- **Role**: Nexus acts as the orchestrator within the system, managing the flow of messages and ensuring that each one
-  is processed according to predefined rules and logic.
-- **Functionality**: It handles various tasks like message validation, transformation, and decision-making on how and
-  where messages should be directed post-processing.
-- **Scalability and Performance**: Designed for high performance and scalability, Nexus can handle a vast volume of
-  messages simultaneously, ensuring minimal latency and high throughput in data processing.
-
-### Conduit
-
-The **Conduit** represents the pathway through which messages, or 'Pulses', are transmitted within the system. It's the
-messenger that ensures the delivery of data from one point to another.
-
-- **Mechanism**: Conduit facilitates the smooth and efficient transport of messages across different parts of the
-  system.
-- **Reliability and Integrity**: Ensuring data integrity, Conduit maintains the fidelity of the messages as they
-  traverse through various processes.
-- **Asynchronous Communication**: It supports asynchronous data flow, allowing for non-blocking and concurrent message
-  transmissions, which is key for a responsive and efficient system.
-
-### Pulse
-
-**Pulse** is the term used to describe the individual units of data or messages that flow through the system.
-
-- **Data Encapsulation**: Each Pulse is a packet of information, encapsulating the necessary data in a well-defined
-  format.
-- **Types and Variability**: Pulses can vary in type and structure, ranging from simple text messages to complex data
-  structures, each tailored to carry specific information relevant to its intended process.
-- **Lifecycle**: The lifecycle of a Pulse includes its creation, transmission through the Conduit, processing in the
-  Nexus, and final delivery or action as dictated by the system's logic.
-
-### PulseFlow
-
-**PulseFlow** is the sophisticated mechanism responsible for handling and manipulating the Pulses as they move through
-the system.
-
-- **Message Handling**: It's specifically designed to process each Pulse, applying necessary transformations, routing,
-  and any other required operations.
-- **Flexibility and Adaptability**: PulseFlow is adept at handling various types of Pulses, capable of adapting its
-  processing logic based on the nature and requirements of each message.
-- **Integration Point**: Acting as a key integration point within the system, it ensures that Pulses are managed
-  efficiently and effectively, readying them for their next phase in the data journey.
