+++ MAINTAINERS.md
@@ -1,0 +1,7 @@
+# Maintainers
+
+The maintainers of this repository are:
+
+| Name | GitHub | Role        |
+| ---- | ------ |-------------|
+| Frank R. Haugen | [frankhaugen](https://github.com/frankhaugen) | Maintainer/Owner |
